FEAT: completed first code review
TODO: When the user is loaded from the server (find the function for thatthe response from the server includes an array of which stories the user has already added to favorites. For each story in the story list, you’ll want to mark it’s favorite to true.e. Then, add two methods to the User class, letting the user favorite or un-favorite a story. These methods will need to take a Story instance, and should change the story be a favorite/un-favorite. They should also send a request to the API so the server knows when a favorite/un-favorite action occurs.